test: 메뉴 추가 인수 테스트 수정
